Move risky position math to effect
